Fix bug and remove backup on startup

+++ controllers/api/bots.js
@@ -263,15 +263,18 @@ router.delete('/lists/:id', (req, res, next) => {
     .exec((err, user) => {
       if (err) return next(err);
       TenThingsList.findOne(req.params.id, (err, list) => {
-        if (config.admins.indexOf(req.auth.userid) >= 0 || req.auth.userid === list.creator) {
-          TenThingsList.findByIdAndRemove(req.params.id, (err, list) => {
-            if (err) return next(err);
-            bot.notifyAdmins('<b>' + list.name + '</b> deleted by ' + user.username);
+        if (err) return next(err);
+        if (list) {
+          if (config.admins.indexOf(req.auth.userid) >= 0 || req.auth.userid === list.creator) {
+            TenThingsList.findByIdAndRemove(req.params.id, (err, list) => {
+              if (err) return next(err);
+              bot.notifyAdmins('<b>' + list.name + '</b> deleted by ' + user.username);
+              res.sendStatus(200);
+            });
+          } else {
+            bot.notifyAdmins(`Unauthorized detletion: <b>${list.name}</b> by ${user.username} (${user._id})`);
             res.sendStatus(200);
-          });
-        } else {
-          bot.notifyAdmins(`Unauthorized detletion: <b>${list.name}</b> by ${user.username} (${user._id})`);
-          res.sendStatus(200);
+          }
         }
       });
     });
